Prep for version 0.3 and break code up into separate files for clarity.
