+++ index.js
@@ -1,80 +1,44 @@
 const versionNo="0.2";
-import {Sections} from './structure.js';
+
 import {unitObj,meas} from './utils.js';
-import {svg,group,rect,hLine,vLine,path,guideRect, pathCmdList,text} from './svg.js';
+export {unitObj,meas};
+import {svg,group,rect,hLine,vLine,path,guideRect, pathCmdList,text,setSvgAttributes} from './svg.js';
+export{path,pathCmdList,guideRect,setSvgAttributes,group,rect,text};
 import {
   LEFT,RIGHT,CENTER,SIDES,
   FOLD,FOLDINNER,FOLDOUT,STRIP,STRIPINNER,EDGE,REFS,
   SLIDE,GUIDE,BASE,SECTIONS,
-  FILL,
+  FILL,STYLE,TEXTANCHOR,
   SIDE,INDEX,REF,X,Y} from './const.js';
-import {Guides,printUsage} from './guides.js';
+export {
+  LEFT,RIGHT,CENTER,SIDES,
+  FOLD,FOLDINNER,FOLDOUT,STRIP,STRIPINNER,EDGE,REFS,
+  SLIDE,GUIDE,BASE,SECTIONS,
+  FILL,STYLE,TEXTANCHOR,
+  SIDE,INDEX,REF,X,Y};
+import {Guides} from './guides.js';
+export {guides}
+import {topCutout,bottomCutout,splitterTabs,addCircleTabs} from './basicCuts.js';
+export {topCutout,bottomCutout,splitterTabs,addCircleTabs};
+import drawSlide from './drawSlide.js';
+import drawGuide from './drawGuide.js';
+import {printUsage,sectionConfig,style,params} from './config.js';
+export {printUsage,params,style};
 
-// PARAMETERS
-var page, stripWidth, finishedWidth, 
-    foldWidth, inner, tapeType, slideLength,
-    panelWidth, baseWidth;
-export var units={
-    x:100,
-    label:"inch",
-    to:function(meas) {
-        return(meas*this.x); 
-    },
-    from:function(meas) {
-      return(meas/this.x);
-    }
-};
-const cutoutFill="#ddd";
-const black="#000";
-const guideStroke="#eee";
-const foldStrokeDash="10,5,3,5";
-const foldStroke="#000";
-const foldStrokeWidth="2";
-const cutStroke="#000";
-const cutStrokeWidth="2";
-
-page = {width: new unitObj(8.5),height: new unitObj(11)};    
-stripWidth = new unitObj(1.25);
-finishedWidth = new unitObj(stripWidth.raw/4+.1);
-foldWidth = new unitObj((stripWidth.raw - 2*finishedWidth.raw)/2);
-inner = new unitObj(.02);
-panelWidth= new unitObj(page.width.raw/3-2*inner.raw);
-baseWidth= new unitObj(page.width.raw-2*panelWidth.raw);
-var hGuideSep = new unitObj(.25);
-var offset = new unitObj(baseWidth.raw*.5);
-export var params = {
-    page:page,
-    stripWidth:stripWidth,
-    finishedWidth:finishedWidth,
-    foldWidth:foldWidth,
-    inner:inner,
-    panelWidth:panelWidth,
-    baseWidth:baseWidth,
-    hGuideSep:hGuideSep,
-    offset:offset,
-    units:units};
-var slideStart = 6;
-var slideEnd = 19;
-var guideEnd = 21;
-var firstTab = 28;
-var secondTab = 35;
-
-export var guides = new Guides();
+var guides = new Guides();
 guides.calculate();
 
-var svgDrawing = svg(page.width.u,page.height.u);
+var svgDrawing = svg(params.page.width.u,params.page.height.u);
 document.body.appendChild(svgDrawing);
 
-var sections = new Sections(params);
-
-//draw the main SVG
-for (let i=0; i<sections.count; i++) {
-  var sect = sections.array[i];
-  var groupObj = group(sect.name);
-  groupObj.setAttribute("transform","translate("+sect.offset.u+" "+0+")");
+for (let sectI in SECTIONS) {
+  var sect=SECTIONS[sectI];
+  var groupObj = group(sect);
+  var sectConfig=sectionConfig.get(sect);
+  groupObj.setAttribute("transform","translate("+sectConfig.offset.u+" "+0+")");
   svgDrawing.appendChild(groupObj);
 
-  var frame = rect(0,0,baseWidth.u,page.height.u,sect.fill);
+  var frame = rect(0,0,params.baseWidth.u,params.page.height.u,sectConfig.fill);
   frame.setAttribute("id",sect.name+"-frame")
   if (sect.name==BASE) setFoldAttributes(frame);
   groupObj.appendChild(frame);
@@ -86,7 +50,7 @@ for (let i=0; i<sections.count; i++) {
       if (guides.ax(X).side(side).ref(ref)) {
         var guide = vLine({x:{SIDE: side,REF:ref}});
         guide.setAttribute("id",(side+"-"+ref));
-        guide.setAttribute("stroke",(guideStroke));
+        guide.setAttribute("stroke",(style.guide.stroke));
         groupObj.append(guide);
       }
     }
@@ -95,230 +59,8 @@ for (let i=0; i<sections.count; i++) {
   if (sect.name==SLIDE) addCircleTabs(groupObj,RIGHT);
 }
 
-function drawGuide(parent) {
-  topCutout(parent,1);
-  topCutout(parent,2.5);
-  topCutout(parent,4);
-
-  var d;
-  d = pathCmdList([
-    {cmd:"M",x:{side:LEFT,ref:FOLDINNER},y:{index:guideEnd}},
-    {cmd:"V",y:{index:slideEnd}},
-    {cmd:"H",x:{side:LEFT,ref:STRIP}},
-    {cmd:"V",y:{index:slideStart}},
-    {cmd:"H",x:{side:LEFT,ref:STRIPINNER}},
-    {cmd:"L",x:{side:LEFT,ref:FOLDINNER},y:{index:slideEnd-6}},
-    {cmd:"V",y:{index:guideEnd}},
-    {cmd:"H",x:{side:RIGHT,ref:FOLDINNER}},
-    {cmd:"V",y:{index:slideEnd-6}},
-    {cmd:"L",x:{side:RIGHT,ref:STRIPINNER},y:{index:slideStart}},
-    {cmd:"H",x:{side:RIGHT,ref:STRIP}},
-    {cmd:"V",y:{index:slideEnd}},
-    {cmd:"H",x:{side:RIGHT,ref:FOLDINNER}},
-    {cmd:"V",y:{index:guideEnd}}, 
-    {cmd:"z"}
-  ]);
-  
-  var cPath = path(d,GUIDE+"-cutout-3");
-  setCutAttributes(cPath);
-  parent.appendChild(cPath);
-  
-  bottomCutout(parent,guideEnd+1,firstTab-guideEnd);
-  bottomCutout(parent,firstTab+2,secondTab-firstTab-1);
-  
-  bottomCutout(parent,secondTab+2,40-secondTab) 
-  bottomCutout(parent,43,.5) 
-}
 drawGuide(svgDrawing.getElementById(GUIDE));
- 
-function drawSlide(parent) {
-  topCutout(parent,1);
-  topCutout(parent,2.5);
-
-  var d = pathCmdList([
-    {cmd:"M",x:{side:LEFT,ref:STRIP},y:{index:slideStart+1}},
-    {cmd:"V",y:{index:slideEnd-4}},
-    {cmd:"H",x:{side:LEFT,ref:FOLDOUT}},
-    {cmd:"V",y:{index:slideEnd}},
-    {cmd:"H",x:{side:LEFT,ref:FOLD}},
-    {cmd:"m",dx:-finishedWidth.u-inner.u,dy:0},
-    {cmd:"v",dy:-2.1*hGuideSep.u},
-    {cmd:"M",x:{side:RIGHT,ref:STRIP},y:{index:slideStart+1}},
-    {cmd:"V",y:{index:slideEnd-4}},
-    {cmd:"H",x:{side:RIGHT,ref:FOLDOUT}},
-    {cmd:"V",y:{index:slideEnd}},
-    {cmd:"H",x:{side:RIGHT,ref:FOLD}},
-    {cmd:"m",dx:finishedWidth.u+inner.u,dy:-4*hGuideSep.u},
-    {cmd:"v",dy:2.1*hGuideSep.u}
-  ]);
-  
-  var bPath = path(d,SLIDE+"-cut-2");
-  setCutAttributes(bPath);
-  bPath.setAttribute("fill","none");
-  parent.appendChild(bPath);
-  d = pathCmdList([
-      {cmd:"M",
-          x:{side:LEFT,ref:FOLD},
-          y:{index:slideStart}},
-      {cmd:"V",y:{index:slideEnd}},
-      {cmd:"M",
-          x:{side:RIGHT,ref:FOLD },
-          y:{index:slideStart}},
-      {cmd:"V",y:{index:slideEnd}},
-  ]);
-  var cPath = path(d,SLIDE+"-fold");
-  setFoldAttributes(cPath);
-  parent.appendChild(cPath);
-  
-  splitterTabs(parent,firstTab);
-  splitterTabs(parent,secondTab);
-  bottomCutout(parent,41.5,.5) 
-  bottomCutout(parent,43,.5) 
-}
 drawSlide(svgDrawing.getElementById(SLIDE));
  
-function setCutAttributes(svgObj) {
-  svgObj.setAttribute("stroke",cutStroke);
-  svgObj.setAttribute("strokeWidth",cutStrokeWidth);
-  svgObj.setAttribute("fill",cutoutFill);
-}
-function setFoldAttributes(svgObj) {
-    svgObj.setAttribute("stroke",foldStroke);
-    svgObj.setAttribute("stroke-width",foldStrokeWidth);
-    svgObj.setAttribute("stroke-dasharray",foldStrokeDash);
-}
-function topCutout(parent,yInd) {
-  var aRect = guideRect({
-    x1:{SIDE:LEFT,REF:STRIP},
-    y1:yInd*hGuideSep.u,
-    w:stripWidth.u,
-    h:.5*hGuideSep.u
-    });
-  setCutAttributes(aRect);
-  aRect.setAttribute("id",parent.name+"-cutout-"+yInd);
-  parent.appendChild(aRect);
-}
-
-function bottomCutout(parent,yInd,hSepCt) {
-  var aRect = guideRect({
-    x1:{SIDE:LEFT,REF:FOLD},
-    y1:yInd*hGuideSep.u,
-    w:2*finishedWidth.u,
-    h:hSepCt*hGuideSep.u
-    });
-  setCutAttributes(aRect);
-  aRect.setAttribute("id",parent.getAttribute("id")+"-cutout-"+yInd);
-  parent.appendChild(aRect);
-}
-
-function splitterTabs(parent,yInd) {
-  var d, cutPath, foldPath;
-  d = pathCmdList([
-    {cmd:"M",x:{side:LEFT, ref:FOLD},y:{index: yInd+3}},
-    {cmd:"H",x:{side:LEFT, ref: FOLDOUT}},
-    {cmd:"V",y:{index:yInd+1}},
-    {cmd:"H",x:{side:LEFT, ref:FOLD}},
-    {cmd:"m",dx:-finishedWidth.u-inner.u,dy:0},
-    {cmd:"v",dy:hGuideSep.u},
-    {cmd:"M",x:{side:RIGHT, ref:FOLD},y:{index:yInd}},
-    {cmd:"H",x:{side: RIGHT, ref: FOLDOUT}},
-    {cmd:"V",y:{index:yInd+2}},
-    {cmd:"H",x:{side:RIGHT, ref: FOLD}},
-    {cmd:"m",dx:finishedWidth.u+inner.u,dy:0},
-    {cmd:"v",dy:-hGuideSep.u}
-  ]);
-  cutPath = path(d,parent.name+"cut-tabs-"+yInd);
-  setCutAttributes(cutPath);
-  cutPath.setAttribute("fill","none");
-  parent.appendChild(cutPath);
-  d = pathCmdList([
-    {cmd:"M",x:{side:LEFT, ref:FOLD },y:{index:yInd+1}},
-    {cmd:"V",y:{index: yInd+3}},
-    {cmd:"M",x:{side:RIGHT, ref:FOLD},y:{index:yInd}},
-    {cmd:"V",y:{index:yInd+2}}
-  ]);
-  foldPath = path(d,parent.name+"fold-tabs-"+yInd);
-  setFoldAttributes(foldPath);
-  parent.appendChild(foldPath);
-}
-
-function addScale(parent) {
-	var inches={};
-	for (let i=0;i<=2;i+=1/16) inches[i]=new unitObj(i);
-	var cms={};
-	for (let i=0;i<=5;i+=.1) cms[Math.round(i*10)/10]=new meas(i/2.54,Math.floor(10*units.x*i/2.54)/10);
-  //console.log(inches);
-  console.log({x1:inches[0].u,x2:inches[2].u,y1:0,y2:hGuideSep.u});
-  //console.log(cms);
-  console.log({x1:cms[0].u,
-    x2:cms[5].u,
-    y1:0,
-    y2:-hGuideSep.u});
-  var scale = group("scale");
-  scale.setAttribute("transform","translate(60,10) rotate(90)");
-  var inchScale=rect(inches[0].u,0,inches[2].u-inches[0].u,hGuideSep.u,"#fff");
-  console.log("inchScale",inchScale);
-  var cmScale=rect(cms[0].u,hGuideSep.u,cms[5].u-cms[0].u,hGuideSep.u,"#fff");
-  inchScale.setAttribute("stroke",black);
-  cmScale.setAttribute("stroke",black);
-  scale.appendChild(inchScale);
-  scale.appendChild(cmScale);
-  var d="";
-  const inchMarkings={1:.5,2:.4,4:.3,8:.2,16:.15};
-  for (let j in inches) {
-    if (j==Math.floor(j)) {
-      scale.appendChild(text(j,j+"-in",{x:inches[j].u,y:.8*hGuideSep.u,fill:"#000",style:"font-size:80%"}));
-    }
-    for (let k in inchMarkings) {
-      if (j*k==Math.floor(j*k)) {//integer
-        d+=pathCmdList([
-          {cmd:"M",x:inches[j].u,y:0},
-          {cmd:"v",dy:inchMarkings[k]*hGuideSep.u}
-        ]);
-        break;
-      }
-    }
-  }
-  const cmMarkings={1:.5,2:.3,10:.2};
-  for (let j in cms) {
-    if (j==Math.floor(j)) {
-      scale.appendChild(text(j,j+"-cm",{x:cms[j].u,y:1.5*hGuideSep.u,fill:"#000",style:"font-size:80%"}));
-    }
-    for (let k in cmMarkings) {
-      if (j*k==Math.floor(j*k)) {//integer
-        d+=pathCmdList([
-          {cmd:"M",x:cms[j].u,y:2*hGuideSep.u},
-          {cmd:"v",dy:-cmMarkings[k]*hGuideSep.u}
-        ]);
-        break;
-      }
-    }
-  }
-  var scalePath = path(d,"scale-markings");
-  scalePath.setAttribute("stroke","#000");
-  parent.appendChild(scale);
-  scale.appendChild(scalePath);
-}
-addScale(svgDrawing);
-
-function addCircleTabs(parent,side) {
-  var sweep = (side==RIGHT) ? 0:1;
-  var xLoc = (side==RIGHT) ? baseWidth.u-10:10;
-  var d, tabHeights={1:2,2:12,3:22,4:32,5:42};
-  for (let tab in tabHeights) {
-    d=pathCmdList([
-      {cmd:"M",x:xLoc,y:{index:tabHeights[tab]}},
-      {cmd:"a",rx:20,ry:20,dx:0,dy:40,sweep:sweep}
-    ]);
-    var circTab=path(d)
-    circTab.setAttribute("id",parent.getAttribute("id")+"-tab-"+tab);
-    setCutAttributes(circTab);
-    circTab.setAttribute(FILL,"none");
-    parent.appendChild(circTab);
-  }
-}
-console.log(panelWidth,baseWidth);
-console.log(baseWidth.u-panelWidth.u)
-
 svgDrawing.appendChild(text("BiasTapeMaker v"+versionNo,"version-tag",{x:"50%",y:20,fill:"#000","text-anchor":"middle"}));
 svgDrawing.appendChild(text("Designed by Kendra Pendolino","credits-kpendolino",{x:"50%",y:40,fill:"#000","text-anchor":"middle"}));
