+++ todo.md
@@ -1,0 +1,23 @@
+## Version 0.3
+# Completed
+* Cleanup guides.js
+
+## Version 0.2
+# TODO
+* Slide
+  * Fold lines - whole length from slideStart to just before last cutouts
+  * get rid of second set of splitter tabs
+  * reduce splitter tab cut to 0.5*hGuideSep
+  * replace slide tabs with regular splitter tab (can I edit the path afterward to move the top cut end point?)
+  * Make slide longer
+  * Add indicator of where to reinforce with tape (slide tabs and lower end of slide)
+* Guide
+  * Sharper angle at the top - angle for maybe only 2*hGuideSep
+  * Widen cutout to make more room for the slide - outset by probably 2x or 3x inner.
+  * shorten 2nd and 3rd bottomCutout to leave a 1.5" long strip for the fabric to go over
+  * Can these cutouts turn into mini guides, with offset slits to allow the fold to come up above?
+* General
+  * Make circle tabs larger (2x)
+  * 4 circle tabs instead of 5
+  * circle tab placement seemed good
+  * Make vertical positioning more adaptive.
