+++ utils.js
@@ -1,13 +1,11 @@
-export { unitObj } ; // to be removed -> replace with rawMeas, uMeas
-export { rawMeas, uMeas, units, debug };
-import {units} from './index.js';
+export { unitObj, rawMeas, uMeas, debug, meas};
+import {units} from './config.js';
 
+//var units=params.units;
 function debug(msg){
     document.body.appendChild(document.createElement("p")).innerHTML+=msg;
 };
-
-
-export function meas(raw,u) {
+function meas(raw,u) {
   this.raw=raw;
   this.u=u;
 }
@@ -19,7 +17,6 @@ function uMeas(u) {
     units.from(u),
     u);
 }
-
 function unitObj(raw) {
   this.raw=raw;
   this.u=Math.round(units.to(raw));
